refactor: refine folder structure
