+++ moqt-core/src/modules/subscription_models.rs
@@ -1,5 +1,3 @@
-pub mod consumer_node;
-pub mod node_registory;
-pub mod producer_node;
+pub mod nodes;
 pub mod subscriptions;
 pub mod tracks;
